refactor(backoffice-v2): moved TCell types to a different file
